add on curve check
