+++ src/smol_curve/constants.rs
@@ -13,3 +13,27 @@ pub const ZETA_TO_TRACE: Fq = Fq::from_montgomery_limbs_64([
     9804873068900332207,
     302335131180501866,
 ]);
+
+/// COEFF_A = -1
+pub const COEFF_A: Fq = Fq::from_montgomery_limbs_64([
+    10157024534604021774,
+    16668528035959406606,
+    5322190058819395602,
+    387181115924875961,
+]);
+
+/// COEFF_D = 3021
+pub const COEFF_D: Fq = Fq::from_montgomery_limbs_64([
+    15008245758212136496,
+    17341409599856531410,
+    648869460136961410,
+    719771289660577536,
+]);
+
+/// -2 COEFF_D / COEFF_A = 6042
+pub const COEFF_K: Fq = Fq::from_montgomery_limbs_64([
+    10844245690243005535,
+    9774967673803681700,
+    12776203677742963460,
+    94262208632981673,
+]);
